@@ -1,16 +1,16 @@
 name: nuke
 description: Super slim state management that gets the job done.
-version: 0.0.8
+version: 0.0.10
 homepage: https://github.com/1N50MN14
 repository: https://github.com/1N50MN14/nuke
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
+  equatable: ^2.0.3
   flutter:
     sdk: flutter
-  equatable: ^1.2.0
   path_to_regexp: ^0.4.0
   uuid: ^3.0.4
 
