@@ -2,8 +2,10 @@
 //  Generated file. Do not edit.
 //
 
+// clang-format off
+
 #include "generated_plugin_registrant.h"
 
 
-void RegisterPlugins(flutter::PluginRegistry* registry) {
+void fl_register_plugins(FlPluginRegistry* registry) {
 }
