@@ -2,12 +2,14 @@
 //  Generated file. Do not edit.
 //
 
+// clang-format off
+
 #ifndef GENERATED_PLUGIN_REGISTRANT_
 #define GENERATED_PLUGIN_REGISTRANT_
 
-#include <flutter/plugin_registry.h>
+#include <flutter_linux/flutter_linux.h>
 
 // Registers Flutter plugins.
-void RegisterPlugins(flutter::PluginRegistry* registry);
+void fl_register_plugins(FlPluginRegistry* registry);
 
 #endif  // GENERATED_PLUGIN_REGISTRANT_
